refactor getActiveValueFromValueTree
https://coveord.atlassian.net/browse/KIT-2668
